refactor(evals): move batch evaluation implementation
